Final version of the project:
- Implemented all CRUD functionality;
- Handled main scenarios of exceptional cases;
- Applied styling to HTML elements, specifically using Bootstrap.

+++ src/main/webapp/js/script.js
@@ -1,47 +1,55 @@
-let playersCount, rowsHtml, currentPage;
-let $rowsLimiter = $('#rowsLimiter'); // Caching the selector for repeated use
+let playersCount, rowsHtml, currentPage,
+    editedElement, isEditedModeEnabled = false;
+
+// Caching the selectors for repeated use
+const $rowsLimiter = $('#rowsLimiter');
+const $playersInfo = $('#playersInfo');
+const $buttonsContainer = $('#buttonsContainer');
+const $saveButton = $('#saveButton');
 
 // Function to parse the current dropdown value
-let parseDropdown = () => parseInt($rowsLimiter.val());
+function parseDropdown() {
+    return parseInt($rowsLimiter.val());
+}
 
 // Function to add player information to the table
-let playersInformationAdder = function (players) {
+function playersInformationAdder(players) {
     rowsHtml = '';
 
-    for (let player of players) {
+    for (const player of players) {
         rowsHtml += `<tr>
                 <td>${player.id}</td>
-                <td>${player.name}</td>
-                <td>${player.title}</td>
-                <td>${player.race}</td>
-                <td>${player.profession}</td>
+                <td class="name">${player.name}</td>
+                <td class="title">${player.title}</td>
+                <td class="race">${player.race}</td>
+                <td class="profession">${player.profession}</td>
                 <td>${player.level}</td>
                 <td>${new Date(player.birthday).toLocaleDateString()}</td>
-                <td>${player.banned}</td>
+                <td class="banned">${player.banned}</td>
                 <td><img src="/img/edit.png" alt="Edit account" data-id="${player.id}" class="editIcon"></td>
                 <td><img src="/img/delete.png" alt="Delete account" data-id="${player.id}" class="deleteIcon"></td>
             </tr>`;
     }
 
-    $('#playersInfo').html(rowsHtml);
+    $playersInfo.html(rowsHtml);
 }
 
 // Function to add pagination buttons
-let buttonsAdder = function (stayOnCurrentPage) {
+function buttonsAdder(stayOnCurrentPage) {
     rowsHtml = 'Pages:';
-    let pagesLimit = Math.ceil(playersCount / parseDropdown());
+    const pagesLimit = Math.ceil(playersCount / parseDropdown());
 
     for (let pageNumber = 1; pageNumber <= pagesLimit; pageNumber++) {
-        rowsHtml += `<button id="page${pageNumber}">${pageNumber}</button>`;
+        rowsHtml += `<button id="page${pageNumber}" class="pageButton">${pageNumber}</button>`;
     }
 
-    $('#buttonsContainer').html(rowsHtml);
+    $buttonsContainer.html(rowsHtml);
 
     if (stayOnCurrentPage)
         updateCurrentPageButton($(`#page${currentPage.textContent}`)[0]);
     else
         updateCurrentPageButton($('#page1')[0]);
-};
+}
 
 // Function to update the active pagination button
 function updateCurrentPageButton(newPage, clickAnotherButton) {
@@ -53,7 +61,7 @@ function updateCurrentPageButton(newPage, clickAnotherButton) {
 }
 
 // Function to handle page change and dropdown value change
-let changePageContext = function(pressedButton, dropdownValueChanged, stayOnCurrentPage) {
+function changePageContext(pressedButton, dropdownValueChanged, stayOnCurrentPage) {
     if (dropdownValueChanged) {
         pressedButton = $('#page1')[0];
         currentPage = pressedButton;
@@ -84,14 +92,15 @@ $.when(
     $.get('/rest/players')
 ).done(function(countResponse, playersResponse) {
     playersCount = countResponse[0];
-    let playersData = playersResponse[0];
+    const playersData = playersResponse[0];
 
     playersInformationAdder(playersData);
     buttonsAdder();
 });
 
 // Event handlers for pagination buttons and dropdown changes using the cached selector
-$('#buttonsContainer').on('click', 'button', function () {
+$buttonsContainer.on('click', 'button', function () {
+    isEditedModeEnabled = false;
     changePageContext(this);
 
     if (currentPage !== this)
@@ -104,7 +113,12 @@ $rowsLimiter.change(function () {
 
 // Event handler for clicking on the delete user icon.
 // Sends a request to delete an account and updates the content on the page
-$('#playersInfo').on('click', '.deleteIcon', function () {
+$playersInfo.on('click', '.deleteIcon', function () {
+    if (isEditedModeEnabled) {
+        showPopup($('#popupDeletionWarning')[0]);
+        return;
+    }
+
     $.ajax({
         url: `/rest/players/${$(this).data('id')}`,
         type: 'DELETE'
@@ -113,3 +127,137 @@ $('#playersInfo').on('click', '.deleteIcon', function () {
         changePageContext(currentPage, false, true);
     })
 });
+
+// Event handler for clicking on the edit user icon.
+// Activates the account editing mode, makes it possible to send data to the server by double-clicking on the icon
+$playersInfo.on('click', '.editIcon', function () {
+    const dataId = $(this).data('id');
+
+    if (!isEditedModeEnabled) {
+        isEditedModeEnabled = true;
+        editedElement = this;
+    } else if (this === editedElement) {
+        const elementsData = [];
+
+        $('.edited').each(function () {
+            elementsData.push($(this).val());
+        });
+
+        updateData(dataId, JSON.stringify({
+            'name': elementsData[0],
+            'title': elementsData[1],
+            'race': elementsData[2].toUpperCase(),
+            'profession': elementsData[3].toUpperCase(),
+            'banned': elementsData[4]
+        }));
+
+        return;
+    } else {
+        showPopup($('#popupEditWarning')[0]);
+        return;
+    }
+
+    this.src = '/img/save.png';
+    $(`.deleteIcon[data-id="${dataId}"]`)[0].style.display = 'none';
+
+    $(this).parent().siblings().html(function (_, oldHtml) {
+
+        // Function that constructs HTML code for dropdown lists
+        const dropDownListBuilder = function (value) {
+            const firstPart = `<option value="${value.toLowerCase()}"`;
+            const secondPart = `>${value}</option>`;
+
+            if (value === oldHtml)
+                return firstPart + ' selected' + secondPart;
+            else
+                return firstPart + secondPart;
+        };
+
+        switch (this.className) {
+            case 'name':
+                return '<label for="editableName" class="visually-hidden">Enter a name to edit:</label>' +
+                `<input id="editableName" class="edited" type="text" minlength="1" maxlength="12" value="${oldHtml}">`;
+            case 'title':
+                return '<label for="editableTitle" class="visually-hidden">Enter title to edit:</label>' +
+                `<input id="editableTitle" class="edited" type="text" minlength="1" maxlength="30" value="${oldHtml}">`;
+            case 'race':
+                return '<label for="editableRace" class="visually-hidden">Select race:</label>' +
+                '<select id="editableRace" class="edited">' +
+                    dropDownListBuilder('HUMAN') +
+                    dropDownListBuilder('DWARF') +
+                    dropDownListBuilder('ELF') +
+                    dropDownListBuilder('GIANT') +
+                    dropDownListBuilder('ORC') +
+                    dropDownListBuilder('TROLL') +
+                    dropDownListBuilder('HOBBIT') +
+                '</select>';
+            case 'profession':
+                return '<label for="editableProfession" class="visually-hidden">Select profession:</label>' +
+                '<select id="editableProfession" class="edited">' +
+                    dropDownListBuilder('WARRIOR') +
+                    dropDownListBuilder('ROGUE') +
+                    dropDownListBuilder('SORCERER') +
+                    dropDownListBuilder('CLERIC') +
+                    dropDownListBuilder('PALADIN') +
+                    dropDownListBuilder('NAZGUL') +
+                    dropDownListBuilder('WARLOCK') +
+                    dropDownListBuilder('DRUID') +
+                '</select>';
+            case 'banned':
+                return '<label for="editableBanInfo" class="visually-hidden">Select ban status:</label>' +
+                '<select id="editableBanInfo" class="edited">' +
+                    dropDownListBuilder('false') +
+                    dropDownListBuilder('true') +
+                '</select>';
+        }
+    });
+});
+
+// A function that sends a POST request to the server and updates the selected account data
+function updateData(id, data) {
+    $.ajax({
+        url: `/rest/players/${id}`,
+        contentType: 'application/json',
+        type: 'POST',
+        data: data
+    }).done(function() {
+        changePageContext(currentPage);
+        isEditedModeEnabled = false;
+    }).fail(function () {
+        showPopup($('#popupEditError')[0]);
+    });
+}
+
+// A function that creates a new account by sending a POST request to the server
+// when the corresponding button is clicked
+$saveButton.click(function () {
+    $.ajax({
+        url: '/rest/players/',
+        contentType: 'application/json',
+        type: 'POST',
+        data: JSON.stringify({
+            'name': $('#setName').val(),
+            'title': $('#setTitle').val(),
+            'race': $('#setRace').val().toUpperCase(),
+            'profession': $('#setProfession').val().toUpperCase(),
+            'birthday': new Date($('#setBirthday').val()).getTime(),
+            'banned': Boolean($('#setBanInfo').val()),
+            'level': Number($('#setLevel').val())
+        })
+    }).done(function () {
+        changePageContext(currentPage);
+        buttonsAdder(true);
+        playersCount++;
+    }).fail(function () {
+        showPopup($('#popupCreateError')[0]);
+    });
+});
+
+// Shows a pop-up window with error or warning information
+function showPopup(element) {
+    element.style.display = 'block';
+
+    setTimeout(function() {
+        element.style.display = 'none';
+    }, 3000);
+}
